# services/java-client/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@example.com
    name: API Support
  description: "API for managing products, shopping carts, warehouse operations, and\
    \ credit card processing"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-commerce API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.example.com/v1
- description: Staging server
  url: https://staging-api.example.com/v1
security:
- ApiKeyAuth: []
- BearerAuth: []
...
paths:
  /products/{productId}:
    get:
      description: Retrieve a product's details using its unique identifier
      operationId: getProduct
      parameters:
      - description: Unique identifier for the product
        explode: false
        in: path
        name: productId
        required: true
        schema:
          format: int32
          minimum: 1
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product found successfully
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Product not found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal server error
      summary: Get product by ID
      tags:
      - Product
      x-accepts:
      - application/json
  /product:
    post:
      description: "Create a product, with the provided details. Server creates the\
        \ new product_id."
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/createProduct_201_response"
          description: Product created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Invalid input data
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal server error
      summary: "Create a product, with the provided details"
      tags:
      - Product
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
    Product:
      example:
        category_id: 456
        some_other_id: 789
        product_id: 12345
        weight: 1250
        sku: ABC123XYZ
        manufacturer: Acme Corporation
      properties:
        product_id:
          description: Unique identifier for the product
          example: 12345
          format: int32
          minimum: 1
          type: integer
        sku:
          description: "Stock Keeping Unit - unique product code. 10 characters from\
            \ the set {A-Z0-9}"
          example: ABC123XYZ
          maxLength: 100
          minLength: 1
          type: string
        manufacturer:
          description: Product manufacturer name
          example: Acme Corporation
          maxLength: 200
          minLength: 1
          type: string
        category_id:
          description: Product category identifier
          example: 456
          format: int32
          minimum: 1
          type: integer
        weight:
          description: Product weight in grams
          example: 1250
          format: int32
          minimum: 0
          type: integer
        some_other_id:
          description: Additional identifier for product
          example: 789
          format: int32
          minimum: 1
          type: integer
      required:
      - category_id
      - manufacturer
      - product_id
      - sku
      - some_other_id
      - weight
      type: object
    Error:
      example:
        details: Product ID must be a positive integer
        error: INVALID_INPUT
        message: The provided input data is invalid
      properties:
        error:
          description: Error code
          example: INVALID_INPUT
          type: string
        message:
          description: Human-readable error message
          example: The provided input data is invalid
          type: string
        details:
          description: Additional error details
          example: Product ID must be a positive integer
          type: string
      required:
      - error
      - message
      type: object
    createProduct_201_response:
      example:
        product_id: 0
      properties:
        product_id:
          description: Unique identifier for the created product
          format: int32
          type: integer
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http